backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
supabase
python-dotenv
httpx
